Add Notification Service: controller, service, repository, gateway route

+++ notification-service/src/main/java/com/medbook/notificationservice/repository/NotificationRepository.java
@@ -5,5 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface NotificationRepository extends JpaRepository<Notification, Long> {
-    List<Notification> findByUserIdOrderByCreatedAtDesc(Integer userId);
+
+    // 🟢 Lấy danh sách thông báo của user, sắp xếp mới nhất trước
+    List<Notification> findByUserIdOrderByCreatedAtDesc(Long userId);
 }
